Move Overlays into its own component file
