+++ src/components/pages/lessons/special-lessons.ts
@@ -1,0 +1,48 @@
+const specialLessons: any = {
+  'javascript-3-ways-to-update-the-content-of-an-array-of-objects-with-javascript':
+    {
+      headline: 'Check out these in-depth courses on JavaScript Arrays',
+      linksTo: [
+        {
+          title: 'Understand JavaScript Arrays',
+          isPro: true,
+          path: 'understand-javascript-arrays',
+          type: 'course',
+          imageUrl:
+            'https://d2eip9sf3oo6c2.cloudfront.net/playlists/square_covers/000/432/714/square_480/EGH_JSarrays.png',
+        },
+        {
+          title: 'Reduce Data with Javascript Array#reduce',
+          isPro: true,
+          path: 'reduce-data-with-javascript-array-reduce',
+          type: 'course',
+          imageUrl:
+            'https://d2eip9sf3oo6c2.cloudfront.net/playlists/square_covers/000/432/557/square_480/EGH_ReduceDataJS.png',
+        },
+      ],
+    },
+
+  'javascript-creating-demo-apis-with-json-server': {
+    headline: 'Build better APIs with these in-depth courses',
+    linksTo: [
+      {
+        title: 'Build a Serverless API with Cloudflare Workers',
+        isPro: false,
+        slug: 'build-a-serverless-api-with-cloudflare-workers-d67ca551',
+        type: 'course',
+        imageUrl:
+          'https://d2eip9sf3oo6c2.cloudfront.net/playlists/square_covers/000/441/045/square_480/EGH_cloudflare-workers_424_2x.png',
+      },
+      {
+        title: 'Building an API with Express',
+        isPro: true,
+        slug: 'building-an-api-with-express-f1ea',
+        type: 'course',
+        imageUrl:
+          'https://d2eip9sf3oo6c2.cloudfront.net/tags/images/000/000/359/square_480/expressjslogo.png',
+      },
+    ],
+  },
+}
+
+export default specialLessons
